refactor: rename `dependencies_container` module to `serve`
