+++ mithril-aggregator/src/dependency_injection/containers/mod.rs
@@ -1,8 +1,8 @@
-mod dependencies_container;
 mod genesis;
+mod serve;
 
-pub use dependencies_container::*;
 pub use genesis::GenesisToolsDependency;
+pub use serve::*;
 
 use std::sync::Arc;
